4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -10,7 +10,10 @@ blogsRouter.get('/', async (request, response) => {
 });
 
 blogsRouter.post('/', async (request, response) => {
-  const { body: { title, author, url, likes }, token } = request;
+  const {
+    body: { title, author, url, likes },
+    token,
+  } = request;
 
   const decodedToken = jwt.verify(token, process.env.SECRET);
   if (!token || !decodedToken.id) {
@@ -29,6 +32,16 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const { token, params: { id } } = request;
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  const blog = await Blog.findById(id);
+
+  if (!token || !decodedToken.id || decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
